Move utils to mechanisms package, nerf the MANIFEST a bit.
