+++ mirgecom/mechanisms/__init__.py
@@ -1,0 +1,64 @@
+"""Utilities for managing thermochemistry mechanisms."""
+
+__copyright__ = """
+Copyright (C) 2020 University of Illinois Board of Trustees
+"""
+
+
+__license__ = """
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+"""
+import sys
+if sys.version_info < (3, 9):
+    # importlib.resources either doesn't exist or lacks the files()
+    # function, so use the PyPI version:
+    import importlib_resources
+else:
+    # importlib.resources has files(), so use that:
+    import importlib.resources as importlib_resources
+
+
+def get_mechanisms_path():
+    """Get central location for thermochemistry mechanism data."""
+    return("mirgecom.mechanisms")
+
+
+def get_mechanism_file_name(mechanism_name):
+    """Form the CTI file name for a mechanism."""
+    return(f"{mechanism_name}.cti")
+
+
+def get_mechanism_path(mechanism_name):
+    """Get the full path to a mechanism file."""
+    mechfile_name = get_mechanism_file_name(mechanism_name)
+    return(f"{get_mechanisms_path()}/{mechfile_name}")
+
+
+def import_mechdata():
+    """Import the mechanism data as a module data resource."""
+    return importlib_resources.files(get_mechanisms_path())
+
+
+def get_mechanism_cti(mechanism_name):
+    """Get the contents of a mechanism CTI file."""
+    mech_data = import_mechdata()
+    mech_file = mech_data / get_mechanism_file_name(mechanism_name)
+    with mech_file.open() as fp:
+        mech_cti = fp.read()
+    return(mech_cti)
